refactor(tokens): move getHoveredColorToken to format
